requirements.txt
replicate
requests
Pillow
python-fasthtml
uvicorn
